+++ src/main/java/net/samtrion/compactdrawers/core/ModConfig.java
@@ -1,18 +1,14 @@
 package net.samtrion.compactdrawers.core;
 
-import com.jaquadro.minecraft.storagedrawers.StorageDrawers;
+import net.minecraftforge.common.config.*;
+import net.samtrion.compactdrawers.CompactDrawers;
 
+@Config(modid = CompactDrawers.MOD_ID, category = "block")
 public final class ModConfig {
-	public static final double StorageFactorCompactDrawer2By1 = 1.5;
-	public static final double StorageFactorCompactDrawer2By1Half = .75;
-	public static final double StorageFactorCompactDrawerHalf = .5;
-	public static final int BaseStorageCompactDrawer;
-
-	static {
-		int baseCompDrawers = StorageDrawers.config.getBlockBaseStorage("compdrawers");
-		if (baseCompDrawers <= 0) {
-			baseCompDrawers = 16;
-		}
-		BaseStorageCompactDrawer = baseCompDrawers;
-	}
+	
+	public static ConfigDrawer drawerHalf = new ConfigDrawer(true, 8);
+	
+	public static ConfigDrawer drawer2By1 = new ConfigDrawer(true, 24);
+	
+	public static ConfigDrawer drawer2By1Half = new ConfigDrawer(true, 12);
 }