+++ src/main/java/net/samtrion/compactdrawers/tile/TileEntityCompactDrawer2By1.java
@@ -5,6 +5,6 @@ import net.samtrion.compactdrawers.core.ModConfig;
 public class TileEntityCompactDrawer2By1 extends TileEntityCompactDrawerBase {
 
 	public TileEntityCompactDrawer2By1() {
-		super(2, "compactdrawers.container.compactdrawer2by1", ModConfig.StorageFactorCompactDrawer2By1);
+		super(2, "compactdrawers.container.compactdrawer2by1", ModConfig.drawer2By1.capacity);
 	}
 }