+++ src/main/java/net/samtrion/compactdrawers/tile/TileEntityCompactDrawer2By1Half.java
@@ -5,6 +5,6 @@ import net.samtrion.compactdrawers.core.ModConfig;
 public class TileEntityCompactDrawer2By1Half extends TileEntityCompactDrawerBase {
 
 	public TileEntityCompactDrawer2By1Half() {
-		super(2, "compactdrawers.container.compactdrawer2by1half", ModConfig.StorageFactorCompactDrawer2By1Half);
+		super(2, "compactdrawers.container.compactdrawer2by1half", ModConfig.drawer2By1Half.capacity);
 	}
 }