+++ src/main/java/net/samtrion/compactdrawers/tile/TileEntityCompactDrawerHalf.java
@@ -5,6 +5,6 @@ import net.samtrion.compactdrawers.core.ModConfig;
 public class TileEntityCompactDrawerHalf extends TileEntityCompactDrawerBase {
 
 	public TileEntityCompactDrawerHalf() {
-		super(3, "compactdrawers.container.compactdrawerhalf", ModConfig.StorageFactorCompactDrawerHalf);
+		super(3, "compactdrawers.container.compactdrawerhalf", ModConfig.drawerHalf.capacity);
 	}
 }